add: collect assigned GPU name if found and show in analyzed view, -> makes ResourceView redundant hence it is removed
